% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pedigree.R
\name{check_pedigree}
\alias{check_pedigree}
\title{Check Pedigree}
\usage{
check_pedigree(ps_ped_file)
}
\arguments{
\item{ps_ped_file}{path to pedigree input file}
}
\value{
list of check results
}
\description{
Simple check of pedigrees
}
